refactored verifier
